DragonFly BSD support (thanks Masu)
git-svn-id: svn://ultimatepp.org/upp/trunk@7895 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Skylark/App.cpp
@@ -12,7 +12,7 @@ namespace Upp {
 
 String GetThreadName()
 {
-	return FormatIntBase(Thread::GetCurrentId() % 33, 35);
+	return FormatIntBase((int64)(Thread::GetCurrentId()) % 33, 35);
 }
 	
 namespace Ini {
